% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_season_stats.R
\name{player_season_stats}
\alias{player_season_stats}
\title{Get Player's Season Stats}
\usage{
player_season_stats(team_id, year, sport)
}
\arguments{
\item{team_id}{Team ID from team_mapping}

\item{year}{Actual Year (example: 2016 or 2015)}

\item{sport}{Tradiational Sport Name or Sport Code}
}
\description{
Returns a dataframe of season stats for all players on the team
in a specific sport and year.
}
\examples{
player_season_stats(697,2016,"MBB")

player_season_stats(697,2015,"Womens Soccer")

}
\keyword{Season}
\keyword{Stats}
\keyword{Team}
